backtask_check/requirements.txt
Django==2.2
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
